# perf(treesitter): cache the parser check in foldexpr