@tomtomjhj 2024-08-28
Personally, I think the better solution is to guide the user to set foldexpr in FileType autocmd, because setting it globally has several issues as explained in the current PR and #28692.

> This caching means that, unlike before this change, buffers without a parser that later get a parser (e.g. the ft was changed (or maybe text was inserted into an empty buffer and then written as a certain ft that has a parser? but would the foldexpr even be set initially in this case...?)).

If you set foldexpr globally, then it would be set in that case. This problem can be fixed by caching with filetype (or language) as key instead of bufnr. However, it would be still broken for the case when you newly install the parser.