@tomtomjhj 2024-08-28
In addition to @lewis6991's suggestion, I recommend adding the following to `b:undo_ftplugin`
* `exe 'lua vim.treesitter.stop()'` for the case when `pcall(vim.treesitter.start, args.buf)` succeeded
* `setl foldexpr< foldmethod<`

Also I set fdm=expr only when fdm is not diff.

relevant code in my config: https://github.com/tomtomjhj/init.vim/blob/6e0a4d2b7354aee5c66f84e7002622702656a5cf/lua/tomtomjhj/treesitter.lua#L48-L56