@GitMurf 2024-08-30
> Note that these were not run with `--clean` but the effect of other plugins is negligible here (and unaffected by these changes).

Nice work on this! This issue has been plaguing me for a while and so happy to finally track it down :) 

Just as fyi, the number of plugins actually exponentially affects the time. From my debugging, it was checking for a parser on every line and looking in every runtime path. And for me atleast, every plugin I have installed added another path it was looking within across every line of the current buffer. So it ended up being N (lines) * P (plugins) * (parser search time).

cc @echasnovski RE: https://github.com/echasnovski/mini.nvim/issues/1151