@ribru17 2024-08-30
@GitMurf thank you for this info! I stand corrected, I'll amend the PR description :+1: 

@tomtomjhj @lewis6991 Thanks for the insight. It is nice to have workarounds but I still feel it is more intuitive to solve this issue in core... till then maybe I can add some documentation encouraging users to set this option in an autocmd rather than globally. I'll also pursue some of the other improvements mentioned in https://github.com/neovim/neovim/issues/28692#issuecomment-2105088651