@GitMurf 2024-08-30
> @tomtomjhj @lewis6991 Thanks for the insight. It is nice to have workarounds but I still feel it is more intuitive to solve this issue in core...

I completely agree especially given that LazyVim has it as its default I believe... or if not default, a very popular / common thing to have treesitter set for folds.

cc @folke as I know he has been on vacation so I'm sure is flooded with notifications and may take a bit before seeing this.